# executor: vectorize hash calculation in hashJoin (#12048)