<!--
Thank you for contributing to TiDB! Please read TiDB's [CONTRIBUTING](https://github.com/pingcap/tidb/blob/master/CONTRIBUTING.md) document **BEFORE** filing this PR.
-->

### What problem does this PR solve? <!--add issue link with summary if exists-->

Fix issue #12048 

### What is changed and how it works?

benchstat result

```
name                                                                    old time/op    new time/op    delta
HashJoinExec/(rows:100000,_concurency:4,_joinKeyIdx:_[0_1])-4              810ms ± 1%     806ms ± 1%   -0.54%  (p=0.021 n=8+8)
HashJoinExec/(rows:100000,_concurency:4,_joinKeyIdx:_[0])-4                145ms ±16%     135ms ± 3%   -7.23%  (p=0.001 n=10+8)
BuildHashTableForList/(rows:100000,_concurency:4,_joinKeyIdx:_[0_1])-4     594ms ± 1%     587ms ± 0%   -1.18%  (p=0.000 n=10+10)
BuildHashTableForList/(rows:100000,_concurency:4,_joinKeyIdx:_[0])-4      13.3ms ± 1%    10.8ms ± 0%  -18.85%  (p=0.000 n=10+9)
BuildHashTableForList/(rows:10,_concurency:4,_joinKeyIdx:_[0])-4          4.17µs ± 4%    4.23µs ± 3%     ~     (p=0.287 n=9+8)

name                                                                    old alloc/op   new alloc/op   delta
HashJoinExec/(rows:100000,_concurency:4,_joinKeyIdx:_[0_1])-4              304MB ± 0%     304MB ± 0%   +0.01%  (p=0.000 n=9+10)
HashJoinExec/(rows:100000,_concurency:4,_joinKeyIdx:_[0])-4                304MB ± 0%     304MB ± 0%   +0.01%  (p=0.000 n=10+10)
BuildHashTableForList/(rows:100000,_concurency:4,_joinKeyIdx:_[0_1])-4    7.11MB ± 0%    7.13MB ± 0%   +0.35%  (p=0.000 n=10+10)
BuildHashTableForList/(rows:100000,_concurency:4,_joinKeyIdx:_[0])-4      7.11MB ± 0%    7.14MB ± 0%   +0.36%  (p=0.000 n=10+10)
BuildHashTableForList/(rows:10,_concurency:4,_joinKeyIdx:_[0])-4          2.00kB ± 0%    2.28kB ± 0%  +13.57%  (p=0.000 n=10+10)

name                                                                    old allocs/op  new allocs/op  delta
HashJoinExec/(rows:100000,_concurency:4,_joinKeyIdx:_[0_1])-4               306k ± 0%      307k ± 0%   +0.35%  (p=0.000 n=10+9)
HashJoinExec/(rows:100000,_concurency:4,_joinKeyIdx:_[0])-4                 306k ± 0%      307k ± 0%   +0.34%  (p=0.000 n=10+10)
BuildHashTableForList/(rows:100000,_concurency:4,_joinKeyIdx:_[0_1])-4     3.79k ± 1%     4.80k ± 2%  +26.63%  (p=0.000 n=8+10)
BuildHashTableForList/(rows:100000,_concurency:4,_joinKeyIdx:_[0])-4       3.79k ± 0%     4.82k ± 0%  +27.04%  (p=0.000 n=10+9)
BuildHashTableForList/(rows:10,_concurency:4,_joinKeyIdx:_[0])-4            16.0 ± 0%      27.0 ± 0%  +68.75%  (p=0.000 n=10+10)
```

```
name                                                                    old time/op    new time/op    delta
HashJoinExec/(rows:100000,_concurency:4,_joinKeyIdx:_[0_1])-4              817ms ± 2%     806ms ± 1%   -1.34%  (p=0.011 n=9+8)
HashJoinExec/(rows:100000,_concurency:4,_joinKeyIdx:_[0])-4                145ms ± 7%     135ms ± 3%   -6.92%  (p=0.000 n=10+8)
BuildHashTableForList/(rows:100000,_concurency:4,_joinKeyIdx:_[0_1])-4     592ms ± 0%     587ms ± 0%   -0.96%  (p=0.000 n=9+10)
BuildHashTableForList/(rows:100000,_concurency:4,_joinKeyIdx:_[0])-4      13.4ms ± 2%    10.8ms ± 0%  -19.19%  (p=0.000 n=9+9)
BuildHashTableForList/(rows:10,_concurency:4,_joinKeyIdx:_[0])-4          4.16µs ± 5%    4.23µs ± 3%     ~     (p=0.173 n=10+8)

name                                                                    old alloc/op   new alloc/op   delta
HashJoinExec/(rows:100000,_concurency:4,_joinKeyIdx:_[0_1])-4              304MB ± 0%     304MB ± 0%   +0.01%  (p=0.000 n=10+10)
HashJoinExec/(rows:100000,_concurency:4,_joinKeyIdx:_[0])-4                304MB ± 0%     304MB ± 0%   +0.01%  (p=0.000 n=10+10)
BuildHashTableForList/(rows:100000,_concurency:4,_joinKeyIdx:_[0_1])-4    7.11MB ± 0%    7.13MB ± 0%   +0.34%  (p=0.000 n=10+10)
BuildHashTableForList/(rows:100000,_concurency:4,_joinKeyIdx:_[0])-4      7.11MB ± 0%    7.14MB ± 0%   +0.37%  (p=0.000 n=10+10)
BuildHashTableForList/(rows:10,_concurency:4,_joinKeyIdx:_[0])-4          2.00kB ± 0%    2.28kB ± 0%  +13.57%  (p=0.000 n=10+10)

name                                                                    old allocs/op  new allocs/op  delta
HashJoinExec/(rows:100000,_concurency:4,_joinKeyIdx:_[0_1])-4               306k ± 0%      307k ± 0%   +0.35%  (p=0.000 n=10+9)
HashJoinExec/(rows:100000,_concurency:4,_joinKeyIdx:_[0])-4                 306k ± 0%      307k ± 0%   +0.34%  (p=0.000 n=10+10)
BuildHashTableForList/(rows:100000,_concurency:4,_joinKeyIdx:_[0_1])-4     3.78k ± 2%     4.80k ± 2%  +26.84%  (p=0.000 n=10+10)
BuildHashTableForList/(rows:100000,_concurency:4,_joinKeyIdx:_[0])-4       3.79k ± 0%     4.82k ± 0%  +27.10%  (p=0.000 n=10+9)
BuildHashTableForList/(rows:10,_concurency:4,_joinKeyIdx:_[0])-4            16.0 ± 0%      27.0 ± 0%  +68.75%  (p=0.000 n=10+10)
```

```
name                                                                    old time/op    new time/op    delta
HashJoinExec/(rows:100000,_concurency:4,_joinKeyIdx:_[0_1])-4              810ms ± 1%     811ms ± 1%     ~     (p=0.541 n=8+9)
HashJoinExec/(rows:100000,_concurency:4,_joinKeyIdx:_[0])-4                145ms ±16%     143ms ±14%     ~     (p=0.739 n=10+10)
BuildHashTableForList/(rows:100000,_concurency:4,_joinKeyIdx:_[0_1])-4     594ms ± 1%     588ms ± 0%   -1.01%  (p=0.000 n=10+9)
BuildHashTableForList/(rows:100000,_concurency:4,_joinKeyIdx:_[0])-4      13.3ms ± 1%    10.9ms ± 2%  -18.17%  (p=0.000 n=10+10)
BuildHashTableForList/(rows:10,_concurency:4,_joinKeyIdx:_[0])-4          4.17µs ± 4%    4.44µs ±10%   +6.52%  (p=0.004 n=9+10)

name                                                                    old alloc/op   new alloc/op   delta
HashJoinExec/(rows:100000,_concurency:4,_joinKeyIdx:_[0_1])-4              304MB ± 0%     304MB ± 0%   +0.01%  (p=0.000 n=9+10)
HashJoinExec/(rows:100000,_concurency:4,_joinKeyIdx:_[0])-4                304MB ± 0%     304MB ± 0%   +0.01%  (p=0.000 n=10+10)
BuildHashTableForList/(rows:100000,_concurency:4,_joinKeyIdx:_[0_1])-4    7.11MB ± 0%    7.14MB ± 0%   +0.41%  (p=0.000 n=10+10)
BuildHashTableForList/(rows:100000,_concurency:4,_joinKeyIdx:_[0])-4      7.11MB ± 0%    7.14MB ± 0%   +0.35%  (p=0.000 n=10+10)
BuildHashTableForList/(rows:10,_concurency:4,_joinKeyIdx:_[0])-4          2.00kB ± 0%    2.28kB ± 0%  +13.57%  (p=0.000 n=10+10)

name                                                                    old allocs/op  new allocs/op  delta
HashJoinExec/(rows:100000,_concurency:4,_joinKeyIdx:_[0_1])-4               306k ± 0%      307k ± 0%   +0.34%  (p=0.000 n=10+10)
HashJoinExec/(rows:100000,_concurency:4,_joinKeyIdx:_[0])-4                 306k ± 0%      307k ± 0%   +0.34%  (p=0.000 n=10+9)
BuildHashTableForList/(rows:100000,_concurency:4,_joinKeyIdx:_[0_1])-4     3.79k ± 1%     4.82k ± 2%  +27.07%  (p=0.000 n=8+10)
BuildHashTableForList/(rows:100000,_concurency:4,_joinKeyIdx:_[0])-4       3.79k ± 0%     4.81k ± 0%  +26.95%  (p=0.000 n=10+10)
BuildHashTableForList/(rows:10,_concurency:4,_joinKeyIdx:_[0])-4            16.0 ± 0%      27.0 ± 0%  +68.75%  (p=0.000 n=10+10)
```

### Check List <!--REMOVE the items that are not applicable-->

Tests <!-- At least one of them must be included. -->

 - Unit test
 - Integration test

Code changes

 - Has exported function/method change

Side effects

 - Possible performance regression

Related changes

<!--
 - Need to cherry-pick to the release branch
 - Need to update the documentation
 - Need to update the `tidb-ansible` repository
-->

Release note