@qw4990 2019-09-10
Could you please investigate why there is performance drawback in the second case `HashJoinExec/(rows:100000,_concurency:4,_joinKeyIdx:_[0])-4                189ms ±25%     215ms ±56%     ~     (p=0.400 n=9+10)`? @sduzh