@sduzh 2019-09-10
> Could you please investigate why there is performance drawback in the second case `HashJoinExec/(rows:100000,_concurency:4,_joinKeyIdx:_[0])-4 189ms ±25% 215ms ±56% ~ (p=0.400 n=9+10)`? @sduzh

I ran the benchmark again and could not reproduce the result.
You can check the latest benchmark result from the pull comment.